manager.py: add support for Python 3
In both Python 2 and 3 the function subprocess.Popen.communicate() returns
binary strings, but in all other parts of the code we use the type 'str' which
is a text type in Python 3. Therefore if necessary return the decoded string so
we do not get for example compilation errors. For decoding try stdout encoding;
if it is not defined, use 'ascii'.